+++ server/src/routes/Auth/index.ts
@@ -1,0 +1,3 @@
+import { FastifyInstance } from "fastify";
+
+export async function authRoutes(fastify: FastifyInstance) {}
