+++ server/src/routes/Game/index.ts
@@ -1,0 +1,3 @@
+import { FastifyInstance } from "fastify";
+
+export async function gameRoutes(fastify: FastifyInstance) {}
