+++ server/src/server.ts
@@ -1,12 +1,10 @@
 import cors from "@fastify/cors";
-import { PrismaClient } from "@prisma/client";
 import Fastify from "fastify";
-import ShortUniqueId from "short-unique-id";
-import { z } from "zod";
-
-const prisma = new PrismaClient({
-  log: ["query"],
-});
+import { authRoutes } from "./routes/Auth";
+import { gameRoutes } from "./routes/Game";
+import { guessRoutes } from "./routes/Guess";
+import { poolRoutes } from "./routes/Pool";
+import { userRoutes } from "./routes/User";
 
 async function bootstrap() {
   const fastify = Fastify({
@@ -17,41 +15,11 @@ async function bootstrap() {
     origin: true,
   });
 
-  fastify.get("/pools/count", async () => {
-    const count = await prisma.pool.count();
-
-    return { count };
-  });
-
-  fastify.get("/users/count", async () => {
-    const count = await prisma.user.count();
-
-    return { count };
-  });
-
-  fastify.get("/guesses/count", async () => {
-    const count = await prisma.guess.count();
-
-    return { count };
-  });
-
-  fastify.post("/pools", async (request, reply) => {
-    const createPoolBody = z.object({
-      title: z.string(),
-    });
-    const { title } = createPoolBody.parse(request.body);
-    const generateShortId = new ShortUniqueId({ length: 6 });
-    const code = String(generateShortId()).toUpperCase();
-
-    await prisma.pool.create({
-      data: {
-        title,
-        code,
-      },
-    });
-
-    return reply.status(201).send({ code });
-  });
+  await fastify.register(poolRoutes);
+  await fastify.register(authRoutes);
+  await fastify.register(gameRoutes);
+  await fastify.register(guessRoutes);
+  await fastify.register(userRoutes);
 
   const port = 3333;
   await fastify.listen({ port: port /*host: "0.0.0.0"*/ });
